+++ web/apps/photos/src/pages/gallery.tsx
@@ -16,6 +16,7 @@ import {
 import { WhatsNew } from "@/new/photos/components/WhatsNew";
 import { shouldShowWhatsNew } from "@/new/photos/services/changelog";
 import type { CollectionSummaries } from "@/new/photos/services/collection/ui";
+import { hasNonSystemCollections } from "@/new/photos/services/collection/ui";
 import downloadManager from "@/new/photos/services/download";
 import {
     getLocalFiles,
@@ -140,7 +141,6 @@ import {
     getDefaultHiddenCollectionIDs,
     getSelectedCollection,
     handleCollectionOps,
-    hasNonSystemCollections,
     splitNormalAndHiddenCollections,
 } from "utils/collection";
 import {
